15-01-19

+++ app/src/main/java/com/mualab/org/user/activity/artist_profile/model/StaffDetailsInfo.java
@@ -33,8 +33,8 @@ public class StaffDetailsInfo {
         public int staffId;
         public String staffName;
         public String job;
-        public int inCallPrice;
-        public int outCallPrice;
+        public double inCallPrice;
+        public double outCallPrice;
         public String completionTime;
         public String staffImage;
         public String bookingType;
